requirements.txt
django==1.8
djangorestframework==3.1.3
django-filter==0.10.0